Refactor task builder naming
