+++ tasks/filemanaged/fmtbuilder/fmtbuilder_test.go
@@ -1,4 +1,4 @@
-package builder
+package fmtbuilder
 
 import (
 	"net/url"
@@ -12,7 +12,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestFileManagedTaskBuilder(t *testing.T) {
+func TestTaskBuilder(t *testing.T) {
 	testCases := []struct {
 		typeName      string
 		path          string
@@ -61,7 +61,7 @@ func TestFileManagedTaskBuilder(t *testing.T) {
 	for _, testCase := range testCases {
 		tc := testCase
 		t.Run(tc.typeName, func(t *testing.T) {
-			taskBuilder := FileManagedTaskBuilder{}
+			taskBuilder := TaskBuilder{}
 			actualTaskI, err := taskBuilder.Build(
 				tc.typeName,
 				tc.path,
