+++ tasks/pkgtask/pkgbuilder/pkgbuilder_test.go
@@ -1,4 +1,4 @@
-package builder
+package pkgbuilder
 
 import (
 	"testing"
@@ -10,7 +10,7 @@ import (
 	"gopkg.in/yaml.v2"
 )
 
-func TestPkgTaskBuilder(t *testing.T) {
+func TestTaskBuilder(t *testing.T) {
 	testCases := []struct {
 		typeName      string
 		path          string
@@ -107,7 +107,7 @@ func TestPkgTaskBuilder(t *testing.T) {
 	for _, testCase := range testCases {
 		tc := testCase
 		t.Run(tc.typeName, func(t *testing.T) {
-			taskBuilder := PkgTaskBuilder{}
+			taskBuilder := TaskBuilder{}
 			actualTaskI, err := taskBuilder.Build(
 				tc.typeName,
 				tc.path,
