+++ tasks/winreg/wrtbuilder/wrtbuilder_test.go
@@ -1,4 +1,4 @@
-package builder
+package wrtbuilder
 
 import (
 	"testing"
@@ -9,7 +9,7 @@ import (
 	"gopkg.in/yaml.v2"
 )
 
-func TestWinRegTaskBuilder(t *testing.T) {
+func TestTaskBuilder(t *testing.T) {
 	testCases := []struct {
 		typeName      string
 		path          string
@@ -167,7 +167,7 @@ func TestWinRegTaskBuilder(t *testing.T) {
 	for _, testCase := range testCases {
 		tc := testCase
 		t.Run(tc.typeName, func(t *testing.T) {
-			taskBuilder := WinRegTaskBuilder{}
+			taskBuilder := TaskBuilder{}
 			actualTaskI, err := taskBuilder.Build(
 				tc.typeName,
 				tc.path,
